requirements.txt
# Core Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0

# Environment and Configuration
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
django-filter==23.3

# Image Processing (for recipe and step images)
Pillow==10.0.1

# CORS handling (for React frontend integration)
django-cors-headers==4.3.1

# Authentication and JWT
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
dj-rest-auth==5.0.2
PyJWT==2.8.0

# OTP and SMS (for mobile verification)
pyotp==2.9.0
twilio==8.10.0

# API Documentation
drf-spectacular==0.26.5

# Development and Testing
pytest==7.4.3
pytest-django==4.5.2

# Utilities
pytz==2023.3
requests==2.31.0
Pillow>=9.0.0
